requirements.txt
python-telegram-bot[job-queue]
sqlalchemy
pandas
openpyxl
matplotlib
xlsxwriter
python-dotenv